pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-sf-utility"
version = "0.2.0"
description = "Metadata-driven file-to-db-to-salesforce data loading utility with real-time web control plane"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.0",
  "PyYAML>=6.0.1",
  "SQLAlchemy>=2.0.0",
  "pydantic>=2.6.0",
  "python-dateutil>=2.9.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
salesforce = ["simple-salesforce>=1.12.6"]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0", "httpx>=0.27.0"]

[project.scripts]
etl-sf-run = "etl_sf.main:main"
etl-sf-web = "etl_sf.web.app:run_web"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
